Using simple Hash tables with many items
For some people it's better to edit values in a single page,
DynamoDB is able to do this, but the previous approach required
one item per configuration, meaning one edit page per configuration.

With the following changes, the user is able to edit many
configurations in a single page using a simple Hash table (with App as
the hash key).

An example is included in the next image:

    http://i.imgur.com/RXoRKCf.png