Implementing loadMetaData
`func loadMetaData(path string, tmp interface{}) {}`

This function will request for the given `path` at the url:
`http://169.254.169.254`
and parse the result according to the `tmp` reflect.TypeOf value.
If it's a struct, then it will do a json.Unmarshal.
If it's a string, then it will just assign the response to that
variable.