Looping over found Scans, ignoring Value
In the current README.md, the database setup requires users to set a hash key
of App and a range key of Value, so, to extend the capacity of configs, users
not only can use just Hash tables with 64KB items with multiple attributes as
configuration variables, but Hash and Range tables with multiple items with
multiple values as configuration variables. The only downside of this approach
is that App and Value keys are ignored, in order to keep the print clean.